% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_exp.R
\name{cluster_exp}
\alias{cluster_exp}
\title{Cluster Samples Based on Gene Expression}
\usage{
cluster_exp(
  exp_data,
  k,
  genes = NULL,
  pca = TRUE,
  n_pcs = 10,
  dist_method = "euclidean",
  hc_method = "complete"
)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing the normalized expression data matrix.}

\item{k}{An integer specifying the number of clusters to generate.}

\item{genes}{A character vector of gene names to be used for clustering. If `NULL`, the top 2000 highly variable genes are selected automatically.}

\item{pca}{Logical. If `TRUE`, principal component analysis (PCA) is performed to reduce dimensionality before clustering. Default is `TRUE`.}

\item{n_pcs}{An integer specifying the number of principal components to retain if `pca = TRUE`. Default is 10.}

\item{dist_method}{A character string specifying the distance metric to use. Can be one of `"euclidean"`, `"manhattan"`, `"pearson"`, or `"spearman"`. Default is `"euclidean"`.}

\item{hc_method}{A character string specifying the agglomeration method for hierarchical clustering. Default is `"complete"`.}
}
\value{
A `SummarizedExperiment` object with an updated column in `colData`, named `"exp_cluster"`, containing the cluster assignments as a factor.
}
\description{
This function performs hierarchical clustering on samples using normalized gene expression data.
}
\details{
If no specific genes are provided, the function automatically selects the top 2000 highly variable genes for clustering.
Optionally, PCA can be applied to reduce dimensionality, which can be useful for datasets with a large number of genes.
}
