% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_gexp.R
\name{filter_gexp}
\alias{filter_gexp}
\title{Filter gene expression data based on counts and sample presence}
\usage{
filter_gexp(exp_data, min_nsamp = 1, min_counts = 10)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing the gene expression data.}

\item{min_nsamp}{An integer. The minimum number of samples in which a gene must have
at least `min_counts` to be retained. Default is 1.}

\item{min_counts}{An integer. The minimum number of counts for a gene to be retained
in each sample. Default is 10.}
}
\value{
A `SummarizedExperiment` object containing only the genes that meet the criteria.
}
\description{
This function filters genes from a `SummarizedExperiment` object based on the
number of counts in at least a certain number of samples. Genes that do not
meet the specified threshold for both the minimum number of samples and the
minimum number of counts per sample are removed.
}
