% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixMarsExcel.R
\name{fixMarsExcel}
\alias{fixMarsExcel}
\title{Fix Excel-Corrupted Gene Names}
\usage{
fixMarsExcel(mat)
}
\arguments{
\item{mat}{A data frame containing a `gene_name` column, possibly with corrupted gene names.}
}
\value{
A data frame with corrected `gene_name` entries.
}
\description{
This function corrects gene names in a data frame that have been
automatically converted by Excel into date formats (e.g., "03/01/01")
and replaces them with their original gene names.
}
