% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_annotation_collection.R
\name{get_annotation_collection}
\alias{get_annotation_collection}
\title{Get Annotation Collections from MSigDB}
\usage{
get_annotation_collection(collections, species = "Homo sapiens")
}
\arguments{
\item{collections}{A character vector of MSigDB sub-collections (e.g., "Hallmark", "CP:KEGG", etc.).}

\item{species}{The species for which to retrieve gene sets (default: "Homo sapiens").}
}
\value{
A data frame with gene set annotations, including collection name, pathway, Ensembl gene ID, and gene symbol.
}
\description{
This function retrieves gene sets from specified MSigDB collections.
}
