% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmptyRows.R
\name{getEmptyRows}
\alias{getEmptyRows}
\title{Get Empty Rows in a Matrix}
\usage{
getEmptyRows(matrix, colToCheck = 1:8)
}
\arguments{
\item{matrix}{A numeric matrix where rows represent observations and columns represent variables.}

\item{colToCheck}{An integer vector specifying which columns to check for zero sums. Default is 1:8.}
}
\value{
An integer vector of row indices where the sum across the specified columns is zero.
}
\description{
This function identifies the rows in a matrix that have a sum of zero across a specified set of columns.
}
