% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gexp_preprocess.R
\name{gexp_preprocess}
\alias{gexp_preprocess}
\title{Pre-process a gene expression matrix to have different gene annotation}
\usage{
gexp_preprocess(
  gexp,
  gene_annotation = NULL,
  keep_annot = "gene_name",
  og_annot = "gene_id",
  keep_stat = sum,
  organism = NULL
)
}
\arguments{
\item{gexp}{A gene expression matrix, with samples in columns and genes in rows}

\item{gene_annotation}{A data.frame of gene annotation, refering to the rows of
`gexp`. It must contain at least two columns: one for the `keep_annot` and one
for the `og_annot`}

\item{keep_annot}{The name of the column in `gene_annotation` that contains
the gene names that you would like to put as rownames of your `gexp`, i.e. the
kept annotation}

\item{og_annot}{The name of the column in `gene_annotation` that contains the
gene names currently being used by `gexp`, i.e. the original annotation.}

\item{keep_stat}{a function for the statistical measure
used for keeping a gene when `keep_annot` doesn't map uniquely to `og_annot`.}
}
\value{
A processed gene expression matrix
}
\description{
Pre-process a gene expression matrix to have different gene annotation
}
