% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highly_variable_genes.R
\name{highly_variable_genes}
\alias{highly_variable_genes}
\title{Identify Highly Variable Genes (HVGs)}
\usage{
highly_variable_genes(gexp, n_hvg = 2000)
}
\arguments{
\item{gexp}{A gene expression matrix or a SummarizedExperiment object containing normalized gene expression data.}

\item{n_hvg}{The number of highly variable genes to return (default is 2000).}
}
\value{
A character vector of gene names representing the highly variable genes.
}
\description{
This function identifies the top n highly variable genes (HVGs) from a gene expression matrix.
It calculates the robust coefficient of variation for each gene and selects the top `n_hvg` genes with the highest variability.
}
