% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gexp.R
\name{normalize_gexp}
\alias{normalize_gexp}
\title{Normalize Gene Expression Data}
\usage{
normalize_gexp(exp_data)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing the gene expression data with a `counts` assay.}
}
\value{
A `SummarizedExperiment` object with an additional `norm` assay containing the normalized expression data.
}
\description{
This function normalizes gene expression counts in a `SummarizedExperiment` object using either `rlog` or `vst` normalization, depending on the number of samples.
}
\details{
The function applies one of two normalization methods:
- `rlog` (regularized log transformation) is used for datasets with fewer than 30 samples.
- `vst` (variance-stabilizing transformation) is used for datasets with 30 or more samples.

The normalized data is stored in a new assay named `"norm"`.
}
