% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_gexp.R
\name{pca_gexp}
\alias{pca_gexp}
\title{Principal Component Analysis (PCA) on Gene Expression Data}
\usage{
pca_gexp(
  exp_data,
  assay = "norm",
  filter = TRUE,
  n_hvg = 2000,
  center = TRUE,
  scale = TRUE
)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing the gene expression data.}

\item{assay}{A character string specifying the assay to use (default is `"norm"` for normalized data).}

\item{filter}{Logical value, whether to filter for highly variable genes (default is `TRUE`).}

\item{n_hvg}{The number of highly variable genes to select if `filter` is `TRUE` (default is 2000).}

\item{center}{Logical value, whether to center the data before performing PCA (default is `TRUE`).}

\item{scale}{Logical value, whether to scale the data before performing PCA (default is `TRUE`).}
}
\value{
A list containing the results of the PCA, including rotation, standard deviation, and scores.
}
\description{
This function performs PCA (Principal Component Analysis) on the gene expression data.
It can optionally filter for highly variable genes (HVGs) before performing PCA.
}
