% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_scatter.R
\name{plot_exp_scatter}
\alias{plot_exp_scatter}
\title{Save or return a scatter plot of two genes from a SummarizedExperiment}
\usage{
plot_exp_scatter(
  exp_data,
  gene1,
  gene2,
  color_var = NA,
  pt_size = 2,
  fname = NULL,
  fwidth = 5,
  fheight = 3
)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing expression data and sample metadata.}

\item{gene1}{A character string indicating the name of the gene for the x-axis.}

\item{gene2}{A character string indicating the name of the gene for the y-axis.}

\item{color_var}{Optional character string indicating a column in metadata used for coloring points. Default is `NA`.}

\item{pt_size}{Numeric. Size of the points in the scatter plot. Default is 2.}

\item{fname}{Optional. A file name (with extension, e.g., `"scatterplot.png"`) to save the plot. If `NULL`, the plot is not saved.}

\item{fwidth}{Numeric. Width of the output figure in inches. Default is 5.}

\item{fheight}{Numeric. Height of the output figure in inches. Default is 3.}
}
\value{
A `ggplot2` object (the scatter plot).
}
\description{
This function extracts expression values for two genes from a `SummarizedExperiment` object,
generates a scatter plot, and optionally saves it to a file.
}
