% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_microenv_heatmap.R
\name{plot_microenv_heatmap}
\alias{plot_microenv_heatmap}
\title{Plot Microenvironment Heatmap}
\usage{
plot_microenv_heatmap(
  exp_data,
  annotations = NA,
  annotation_prop = 0.1,
  annotation_colors = NULL,
  fname = NULL,
  fwidth = 7,
  fheight = 5,
  ...
)
}
\arguments{
\item{exp_data}{A SummarizedExperiment object containing microenvironment population scores in its metadata.}

\item{annotations}{Optional. Sample annotations to include in the heatmap. Defaults to NA.}

\item{annotation_prop}{A numeric value specifying the proportion of the heatmap allocated to annotations. Defaults to 0.1.}

\item{annotation_colors}{Optional. A list of colors for annotations. Defaults to NULL.}

\item{fname}{Optional. A character string specifying the file name to save the heatmap. Defaults to NULL.}

\item{fwidth}{Numeric. The width of the saved heatmap file. Defaults to 7.}

\item{fheight}{Numeric. The height of the saved heatmap file. Defaults to 5.}

\item{...}{Additional parameters passed to the heatmap plotting function.}
}
\value{
A heatmap plot object.
}
\description{
This function generates a heatmap of microenvironment population scores, with optional sample annotations.
}
