% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_path_scatter.R
\name{plot_path_scatter}
\alias{plot_path_scatter}
\title{Plot a scatter plot of two pathway scores}
\usage{
plot_path_scatter(
  exp_data,
  pathway1,
  pathway2,
  color_var = NA,
  pt_size = 2,
  fname = NULL,
  fwidth = 5,
  fheight = 3
)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing the pathway scores in its metadata.}

\item{pathway1}{A character string indicating the name of the first pathway to plot.}

\item{pathway2}{A character string indicating the name of the second pathway to plot.}

\item{color_var}{A character string indicating the column in the metadata for coloring the points. Default is `NA` (no coloring).}

\item{pt_size}{Numeric. The size of the points in the scatter plot. Default is 2.}

\item{fname}{Optional. A file name to save the plot (e.g., `"scatterplot.png"`). If `NULL`, the plot will not be saved.}

\item{fwidth}{Numeric. The width of the output figure in inches. Default is 5.}

\item{fheight}{Numeric. The height of the output figure in inches. Default is 3.}
}
\value{
A `ggplot2` object containing the scatter plot.
}
\description{
This function generates a scatter plot of two selected pathway scores extracted from
a `SummarizedExperiment` object. It uses the `pathway_scores` metadata and
optionally colors the points based on a variable from the sample annotations.
}
