% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_boxplot.R
\name{plot_pathway_boxplot}
\alias{plot_pathway_boxplot}
\title{Plot Pathway Boxplot}
\usage{
plot_pathway_boxplot(
  exp_data,
  pathway,
  annotation,
  color_var = NA,
  pt_size = 1,
  summary_type = c("choose", "line", "box")[1],
  stat_comparisons = NA,
  stat_format = NULL,
  fname = NULL,
  fwidth = 5,
  fheight = 3
)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing pathway scores in the `metadata(exp_data)[["pathway_scores"]]` slot.}

\item{pathway}{A character string specifying the name of the pathway to plot.}

\item{annotation}{A character string specifying the column in `colData(exp_data)` to use for grouping samples in the plot.}

\item{color_var}{A character string specifying a column in `colData(exp_data)` to use for coloring the plot. Default is `NA` (no coloring).}

\item{pt_size}{Numeric; size of the points in the plot. Default is `1`.}

\item{summary_type}{Character; type of summary to overlay on the plot. Options are `"choose"` (default), `"line"`, or `"box"`.}

\item{stat_comparisons}{A list of character vectors specifying groups for pairwise statistical comparisons. Default is `NA` (no statistical comparisons).}

\item{stat_format}{A character string specifying the format for displaying statistical results. Default is `NULL` (no formatting).}

\item{fname}{A character string specifying the file name to save the plot. Default is `NULL` (do not save).}

\item{fwidth}{Numeric; width of the saved plot in inches. Default is `5`.}

\item{fheight}{Numeric; height of the saved plot in inches. Default is `3`.}
}
\value{
A ggplot object representing the boxplot.
}
\description{
Creates a boxplot to visualize pathway scores across annotations, with optional statistical comparisons and customization.
}
\details{
The function extracts pathway scores from the `metadata(exp_data)[["pathway_scores"]]` slot and merges them with sample annotations from `colData(exp_data)`.
It generates a boxplot using a helper function `.plt_boxplot`. Statistical comparisons can be added to the plot, and the plot can be saved as an image file if `fname` is provided.

The `summary_type` argument determines the type of summary overlay:
\describe{
  \item{"choose"}{No additional summary overlay.}
  \item{"line"}{Adds a line connecting the median of each group.}
  \item{"box"}{Adds a summary box around each group.}
}
}
