% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_heatmap.R
\name{plot_pathway_heatmap}
\alias{plot_pathway_heatmap}
\title{Plot Pathway Heatmap}
\usage{
plot_pathway_heatmap(
  pathway_scores,
  pathways = 1:20,
  annotations = NA,
  annotation_prop = 0.1,
  annotation_colors = NULL,
  fname = NULL,
  fwidth = 7,
  fheight = 5,
  ...
)
}
\arguments{
\item{pathway_scores}{Either a data frame of pathway scores or a SummarizedExperiment object
containing pathway scores in its metadata.}

\item{pathways}{A vector specifying the pathways to include in the heatmap. Defaults to the first 20 pathways.}

\item{annotations}{Optional. Sample annotations to include in the heatmap. Defaults to NA.}

\item{annotation_prop}{A numeric value specifying the proportion of the heatmap allocated to annotations. Defaults to 0.1.}

\item{annotation_colors}{Optional. A list of colors for annotations. Defaults to NULL.}

\item{fname}{Optional. A character string specifying the file name to save the heatmap. Defaults to NULL.}

\item{fwidth}{Numeric. The width of the saved heatmap file. Defaults to 7.}

\item{fheight}{Numeric. The height of the saved heatmap file. Defaults to 5.}

\item{...}{Additional parameters passed to the heatmap plotting function.}
}
\value{
A heatmap plot object.
}
\description{
This function generates a heatmap of pathway scores, with optional sample annotations.
}
