% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progeny_heatmap.R
\name{plot_progeny_heatmap}
\alias{plot_progeny_heatmap}
\title{Plot Progeny Heatmap}
\usage{
plot_progeny_heatmap(
  progeny_scores,
  annotations = NA,
  annotation_prop = 0.1,
  annotation_colors = NULL,
  fname = NULL,
  fwidth = 7,
  fheight = 5,
  ...
)
}
\arguments{
\item{progeny_scores}{A data frame of pathway scores or a `SummarizedExperiment` object containing pathway scores in its metadata.}

\item{annotations}{(Optional) Annotations for the samples to be displayed as tracks on the heatmap. Default is `NA`, meaning no annotations are used.}

\item{annotation_prop}{Proportion of the heatmap width allocated to annotations, if provided. Default is `0.1`.}

\item{annotation_colors}{(Optional) A named list of colors for the annotation tracks. Names must match the annotation variables.}

\item{fname}{(Optional) File name to save the heatmap as an image. If `NULL`, the heatmap is displayed but not saved. Default is `NULL`.}

\item{fwidth}{Width of the saved heatmap image in inches. Default is `7`.}

\item{fheight}{Height of the saved heatmap image in inches. Default is `5`.}

\item{...}{Additional arguments passed to the internal heatmap plotting function.}
}
\value{
A heatmap object showing Progeny pathway scores.
}
\description{
Generates a heatmap for Progeny pathway scores, with optional sample annotations and customization options.
}
\details{
If `progeny_scores` is a data frame, it is assumed to contain pathway scores with row names as pathway names and column names as sample IDs.
If it is a `SummarizedExperiment` object, pathway scores are extracted from the metadata under `progeny_scores`.

Annotations can be added to the heatmap if provided, and their appearance can be customized using `annotation_colors`.
}
