% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Plot a Venn diagram for two or three vectors}
\usage{
plot_venn(
  v1,
  v2,
  universe_size = NULL,
  v1_name = "V1",
  v2_name = "V2",
  v3 = NULL,
  v3_name = "V3",
  fills = TRUE,
  quantities = TRUE,
  title = NULL,
  ...
)
}
\arguments{
\item{v1}{A vector of elements in set 1.}

\item{v2}{A vector of elements in set 2.}

\item{universe_size}{Optional total size of the universe for Fisher's test (2-set only).}

\item{v1_name}{Name for set 1.}

\item{v2_name}{Name for set 2.}

\item{v3}{Optional vector of elements in set 3 for a 3-way Venn diagram.}

\item{v3_name}{Name for set 3.}

\item{fills}{Colors to fill the diagram.}

\item{quantities}{Whether to display quantities.}

\item{title}{Title of the plot.}

\item{...}{Additional parameters passed to `plot()`.}
}
\value{
A ggplot object representing the Venn diagram.
}
\description{
This function plots a Venn diagram using the `euler` function from the eulerr package.
}
