% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_cv.R
\name{robust_cv}
\alias{robust_cv}
\title{Robust Coefficient of Variation}
\usage{
robust_cv(v)
}
\arguments{
\item{v}{A numeric vector of values for which the robust coefficient of variation
is calculated.}
}
\value{
A numeric value representing the robust coefficient of variation for the input vector.
}
\description{
This function computes the robust coefficient of variation for a given vector.
The robust coefficient of variation is calculated as the median absolute deviation
(MAD) divided by the median of the data, providing a measure of variability
that is less sensitive to outliers.
}
\examples{
robust_cv(c(1, 2, 3, 4, 5))
robust_cv(c(1, 1, 1, 1, 1))

}
