% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_metadata.R
\name{cluster_metadata}
\alias{cluster_metadata}
\title{Cluster Metadata in Expression Data}
\usage{
cluster_metadata(
  exp_data,
  k,
  metadata_name = "pathway_scores",
  pca = TRUE,
  n_pcs = 10,
  features = NULL,
  dist_method = "euclidean",
  hc_method = "complete"
)
}
\arguments{
\item{exp_data}{A SummarizedExperiment object containing expression data and metadata.}

\item{k}{An integer specifying the number of clusters to generate.}

\item{metadata_name}{A character string specifying the name of the metadata variable
to use for clustering. Defaults to "pathway_scores".}

\item{pca}{Logical. If TRUE, principal component analysis (PCA) is applied before clustering. Defaults to TRUE.}

\item{n_pcs}{An integer specifying the number of principal components to retain if PCA is applied. Defaults to 10.}

\item{features}{A character vector of feature names to use for clustering. If NULL, all features are used. Defaults to NULL.}

\item{dist_method}{A character string specifying the distance metric to use for hierarchical clustering. Defaults to "euclidean".}

\item{hc_method}{A character string specifying the hierarchical clustering linkage method. Defaults to "complete".}
}
\value{
A SummarizedExperiment object with updated sample annotations including the clustering results in the "path_cluster" column.
}
\description{
This function performs hierarchical clustering on specified metadata scores
within the provided expression dataset. It updates the dataset's sample annotations
with the clustering results.
}
\details{
If no features are provided, all rows in the specified metadata are used for clustering.
  PCA can be applied to reduce dimensionality before clustering.
}
