% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pathway_genes.R
\name{get_pathway_genes}
\alias{get_pathway_genes}
\title{Get Pathway Genes}
\usage{
get_pathway_genes(
  annotation_collection,
  pathway_names,
  annotation = "gene_symbol"
)
}
\arguments{
\item{annotation_collection}{A data frame containing pathway annotations. It should have at least two columns: `pathway` (pathway names) and a column for gene annotations (e.g., `gene_symbol` or `gene_id`).}

\item{pathway_names}{A character vector specifying the names of pathways for which to retrieve associated genes.}

\item{annotation}{A character string specifying the column in `annotation_collection` to use for gene annotations. Default is `"gene_symbol"`.}
}
\value{
A vector of genes associated with the specified pathways. The type of genes (e.g., symbols or Ensembl IDs) depends on the `annotation` argument.
}
\description{
Extracts genes associated with specific pathways from an annotation collection.
}
\details{
This function filters the provided annotation collection for the specified pathway names and extracts the corresponding genes based on the selected annotation column.

The `annotation` argument allows for flexibility in retrieving genes by different identifiers, such as `gene_symbol` or `gene_id`.
}
