% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_counter.R
\name{mcp_counter}
\alias{mcp_counter}
\title{MCP-counter wrapper for SummarizedExperiment objects}
\usage{
mcp_counter(exp_data, species, feature_type = "gene_name")
}
\arguments{
\item{exp_data}{A SummarizedExperiment object containing TPM expression data in the "tpm" assay.}

\item{species}{Character. Either "Homo sapiens" or "Mus musculus".}

\item{feature_type}{Character. Either "gene_name" or "ensembl_gene_id".}
}
\value{
A data.frame with MCPcounter scores.
}
\description{
Estimates immune and stromal cell populations using MCPcounter or mMCPcounter.
}
