% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcp_counter.R
\name{mcp_counter}
\alias{mcp_counter}
\title{Estimate Immune and Stromal Cell Populations using MCPcounter}
\usage{
mcp_counter(exp_data, species, feature_type = "gene_name")
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing normalized gene expression data.}

\item{species}{Character. Either "Homo sapiens" or "Mus musculus", specifying the organism.}

\item{feature_type}{Character. Either "gene_name" or "ensembl_gene_id", specifying the gene annotation used in rownames.}
}
\value{
A data frame with estimated cell population scores for each sample.
}
\description{
This function estimates the abundance of various immune and stromal cell populations
in bulk transcriptomic data using the MCPcounter algorithm (for human) or mMCPcounter (for mouse).
}
