% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_gexp.R
\name{normalize_gexp}
\alias{normalize_gexp}
\title{Normalize Gene Expression Data}
\usage{
normalize_gexp(exp_data)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing the gene expression counts
in the `counts` assay.}
}
\value{
A `SummarizedExperiment` object with the normalized data stored in the `norm` assay.
}
\description{
This function normalizes gene expression data using either the `rlog` or `vst` method,
depending on the number of samples. For fewer than 30 samples, `rlog` normalization is
applied; for 30 or more samples, `vst` normalization is used.
}
