% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathwayFGSEA.R
\name{pathwayFGSEA}
\alias{pathwayFGSEA}
\title{Pathway Enrichment Using FGSEA}
\usage{
pathwayFGSEA(diffexp, pathwayCollection)
}
\arguments{
\item{diffexp}{A data frame of differential expression results, with row names as gene IDs
and a `log2FoldChange` column used for ranking genes.}

\item{pathwayCollection}{A data frame with at least two columns: one for pathway names (`pathway`)
and one for gene symbols (`gene_symbol`).}
}
\value{
A data frame returned by `fgseaMultilevel()`, including columns such as:
\describe{
  \item{pathway}{Name of the pathway}
  \item{pval}{P-value of enrichment}
  \item{padj}{Adjusted p-value (FDR)}
  \item{ES}{Enrichment Score}
  \item{NES}{Normalized Enrichment Score}
  \item{leadingEdge}{Vector of leading-edge genes}
}
}
\description{
Performs fast gene set enrichment analysis (FGSEA) using the `fgseaMultilevel` method,
based on a ranked list of differentially expressed genes and a collection of pathways.
}
