% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_boxplot.R
\name{plot_exp_boxplot}
\alias{plot_exp_boxplot}
\title{Plot Expression Boxplot}
\usage{
plot_exp_boxplot(
  exp_data,
  gene,
  annotation,
  color_var = NA,
  pt_size = 1,
  summary_type = c("choose", "line", "box")[1],
  stat_comparisons = NA,
  stat_format = NULL,
  fname = NULL,
  fwidth = 5,
  fheight = 3
)
}
\arguments{
\item{exp_data}{A data object containing expression data.}

\item{gene}{A character string specifying the gene for which expression is plotted.}

\item{annotation}{A character string specifying the grouping variable for the boxplot.}

\item{color_var}{Optional. A character string specifying the variable to color the points. Defaults to NA.}

\item{pt_size}{Numeric. The size of the points in the boxplot. Defaults to 1.}

\item{summary_type}{A character string specifying the type of summary to display ("choose", "line", or "box"). Defaults to "choose".}

\item{stat_comparisons}{Optional. Statistical comparisons to be displayed on the plot. Defaults to NA.}

\item{stat_format}{Optional. A format for statistical annotations. Defaults to NULL.}

\item{fname}{Optional. A character string specifying the file name to save the plot. Defaults to NULL.}

\item{fwidth}{Numeric. The width of the saved plot file. Defaults to 5.}

\item{fheight}{Numeric. The height of the saved plot file. Defaults to 3.}
}
\value{
A ggplot object representing the boxplot.
}
\description{
This function generates a boxplot for the expression of a specified gene, grouped by a given annotation.
}
