% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_heatmap.R
\name{plot_exp_heatmap}
\alias{plot_exp_heatmap}
\title{Plot Heatmap of Gene Expression with Optional Annotations}
\usage{
plot_exp_heatmap(
  expData,
  genes,
  annotations = NA,
  assay = "norm",
  gene_name = "gene_name",
  annotation_prop = 0.1,
  annotation_colors = NULL,
  fname = NULL,
  fwidth = 7,
  fheight = 5,
  ...
)
}
\arguments{
\item{expData}{A `SummarizedExperiment` object containing the expression data. The `norm` assay is used by default.}

\item{genes}{A character vector of gene names or IDs to include in the heatmap.}

\item{annotations}{A data frame or list of annotations to overlay on the heatmap. Default is `NA` (no annotations).}

\item{assay}{The assay to use from the `expData` object. Default is `"norm"`.}

\item{gene_name}{The column name in `rowData(expData)` that contains the gene identifiers. Default is `"gene_name"`.}

\item{annotation_prop}{The proportion of the heatmap's height to allocate to the annotation tracks. Default is 0.1.}

\item{annotation_colors}{A vector of colors for the annotation tracks. Default is `NULL`.}

\item{fname}{Optional. The file name to save the heatmap plot to. Default is `NULL` (no file is saved).}

\item{fwidth}{The width (in inches) of the saved heatmap. Default is 7 inches.}

\item{fheight}{The height (in inches) of the saved heatmap. Default is 5 inches.}

\item{...}{Additional arguments to pass to the `plt_heatmap` function.}
}
\value{
A `ggplot` object representing the heatmap.
}
\description{
This function generates a heatmap of gene expression data from a `SummarizedExperiment` object,
with the option to add annotations to the heatmap. The data is pre-processed before plotting,
and the heatmap can be customized in terms of colors, font size, and file output.
}
