% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_scatter.R
\name{plot_exp_scatter}
\alias{plot_exp_scatter}
\title{Plot Expression Scatter Plot}
\usage{
plot_exp_scatter(
  exp_data,
  gene1,
  gene2,
  color_var = NA,
  pt_size = 2,
  fname = NULL,
  fwidth = 5,
  fheight = 3
)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing gene expression data in the assays slot.}

\item{gene1}{A character string specifying the first gene for the x-axis.}

\item{gene2}{A character string specifying the second gene for the y-axis.}

\item{color_var}{A character string specifying a column in `colData(exp_data)` to use for coloring the points. Default is `NA` (no coloring).}

\item{pt_size}{Numeric; size of the points in the scatter plot. Default is `2`.}

\item{fname}{A character string specifying the file name to save the plot. Default is `NULL` (do not save).}

\item{fwidth}{Numeric; width of the saved plot in inches. Default is `5`.}

\item{fheight}{Numeric; height of the saved plot in inches. Default is `3`.}
}
\value{
A ggplot object representing the scatter plot.
}
\description{
Creates a scatter plot to visualize the expression relationship between two genes, with optional coloring by a variable.
}
