% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_volcano.R
\name{plot_exp_volcano}
\alias{plot_exp_volcano}
\title{Plot Volcano Plot of Differential Expression Results}
\usage{
plot_exp_volcano(
  diffexp,
  nb = 10,
  title = "Volcano Plot of Differential Expression",
  lfc_threshold = c(-1, 1),
  color_up = "#0072B2",
  color_down = "#D32F2F",
  color_ns = "gray80",
  fname = NULL
)
}
\arguments{
\item{diffexp}{A data frame containing differential expression results. Must include:
- `log2FoldChange`: The log2 fold change values for each gene.
- `padj`: The adjusted p-value for each gene.}

\item{nb}{The number of genes that have an annotation}

\item{lfc_threshold}{A numeric vector of length 2 giving the log2 fold change}

\item{color_up}{Color used for upregulated genes (default is "#0072B2").}

\item{color_down}{Color used for downregulated genes (default is "#D55E00").}

\item{color_ns}{Color used for non-significant genes (default is "gray80").}
}
\value{
A `ggplot` object representing the volcano plot.
}
\description{
Generates a volcano plot based on the results of differential expression analysis,
highlighting upregulated and downregulated genes, with labels for top significant genes.
}
