% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_volcano.R
\name{plot_exp_volcano}
\alias{plot_exp_volcano}
\title{Plot Volcano Plot of Differential Expression Results}
\usage{
plot_exp_volcano(diffexp)
}
\arguments{
\item{diffexp}{A data frame containing differential expression results. Must include:
- `log2FoldChange`: The log2 fold change values for each gene.
- `padj`: The adjusted p-value for each gene.}
}
\value{
A `ggplot` object representing the volcano plot.
}
\description{
Generates a volcano plot based on the results of differential expression analysis,
highlighting upregulated and downregulated genes, with labels for top significant genes.
}
