% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway_heatmap.R
\name{plot_pathway_heatmap}
\alias{plot_pathway_heatmap}
\title{Plot Heatmap of Pathway Scores with Optional Annotations}
\usage{
plot_pathway_heatmap(
  pathway_scores,
  pathways = 1:20,
  annotations = NA,
  annotation_prop = 0.1,
  annotation_colors = NULL,
  fname = NULL,
  fwidth = 7,
  fheight = 5,
  ...
)
}
\arguments{
\item{pathway_scores}{A `data.frame` or a `SummarizedExperiment` object. If a `data.frame`, it should contain pathway scores. If a `SummarizedExperiment`, it should have pathway scores stored in the metadata.}

\item{pathways}{A vector of indices or pathway names to be included in the heatmap. Default is 1:20.}

\item{annotations}{Optional. A data frame or list of sample annotations to overlay on the heatmap. Default is `NA` (no annotations).}

\item{annotation_prop}{Proportion of the heatmap's height to allocate to the annotation tracks. Default is 0.1.}

\item{annotation_colors}{A vector of colors to use for the annotation tracks. Default is `NULL`.}

\item{fname}{Optional. The file name to save the heatmap plot to. Default is `NULL` (no file is saved).}

\item{fwidth}{The width (in inches) of the saved heatmap. Default is 7 inches.}

\item{fheight}{The height (in inches) of the saved heatmap. Default is 5 inches.}

\item{...}{Additional arguments to pass to the `plt_heatmap` function.}
}
\value{
A `ggplot` object representing the heatmap.
}
\description{
This function generates a heatmap of pathway scores from either a data frame or a `SummarizedExperiment`
object containing pathway scores. Annotations can optionally be added to the heatmap for further customization.
}
