% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{PCA Plot for Gene Expression Data}
\usage{
plot_pca(
  exp_data,
  color = NA,
  shape = NA,
  fname = "results/qc/plot_PCA.pdf",
  pcs = c(1, 2),
  res_name = "pca_res",
  id_name = "sample_id",
  point_size = 2,
  color_in = c("color", "fill")[1],
  out = c("plotly", "ggplot")[1]
)
}
\arguments{
\item{exp_data}{A SummarizedExperiment object containing gene expression data.}

\item{color}{Variable in the data for coloring the points (default is NA).}

\item{shape}{Variable in the data for changing the shape of the points (default is NA).}

\item{fname}{File name to save the plot (default is "results/qc/plot_PCA.pdf").}

\item{pcs}{A vector of two integers specifying which principal components to plot (default is c(1, 2)).}

\item{res_name}{Name of the PCA result stored in the metadata of the `exp_data` object (default is "pca_res").}

\item{id_name}{Column name for sample identifiers (default is "sample_id").}

\item{point_size}{Size of the points in the plot (default is 2).}

\item{color_in}{Either "color" or "fill" specifying how to handle color mapping (default is "color").}

\item{out}{Output format: either "plotly" for interactive or "ggplot" for static (default is "plotly").}
}
\value{
A ggplot or plotly object depending on the output format selected.
}
\description{
PCA Plot for Gene Expression Data
}
