% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc_filters.R
\name{plot_qc_filters}
\alias{plot_qc_filters}
\title{Plot Quality Control (QC) Filters for Gene Expression Data}
\usage{
plot_qc_filters(
  gexp,
  fname = "results/qc/plot_QC_filters.pdf",
  min_nfeats = 5000,
  min_ncounts = 5e+06,
  out = c("plotly", "ggplot")[1]
)
}
\arguments{
\item{gexp}{A `SummarizedExperiment` object containing gene expression data.}

\item{fname}{The file name to save the plot (default is `"results/qc/plot_QC_filters.pdf"`).}

\item{min_nfeats}{Minimum number of features (genes) required to pass QC (default is 5000).}

\item{min_ncounts}{Minimum number of counts required to pass QC (default is 5e6).}

\item{out}{Output format for the plot: `"plotly"` or `"ggplot"` (default is `"plotly"`).}
}
\value{
A plotly or ggplot object based on the output format specified.
}
\description{
This function creates a plot visualizing the quality control (QC) status of samples based on the number of genes
and counts. It allows for filtering and visualizing samples that pass or fail QC thresholds.
}
