% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_exp_hm.R
\name{prep_exp_hm}
\alias{prep_exp_hm}
\title{Prepare Gene Expression Data for Heatmap Visualization}
\usage{
prep_exp_hm(expData, genes, assay = "norm", gene_name = "gene_name")
}
\arguments{
\item{expData}{A `SummarizedExperiment` object.}

\item{genes}{Character vector of gene names or IDs to extract.}

\item{assay}{Character. The name of the assay to use (default is "norm").}

\item{gene_name}{Character. The column name in `rowData` containing gene names. Default is "gene_name".}
}
\value{
A list containing:
\describe{
  \item{table}{A data frame with expression values and sample annotations.}
  \item{colv}{Column variable to use as the x-axis (typically sample IDs).}
  \item{rowv}{Row variable(s) corresponding to the genes.}
}
}
\description{
This function extracts and formats gene expression data from a `SummarizedExperiment`
object for heatmap plotting. It returns a long-format data frame along with identifiers for samples and features.
}
