% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebase_gexp.R
\name{rebase_gexp}
\alias{rebase_gexp}
\title{Rebase gene expression matrix using a specific gene annotation}
\usage{
rebase_gexp(exp_data, annotation = "gene_name")
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing the gene expression data.}

\item{annotation}{A character string specifying the annotation to use as the primary annotation (default is "gene_name").}
}
\value{
A `SummarizedExperiment` object with updated gene expression and annotations.
}
\description{
This function rebases the gene expression matrix using a selected annotation,
and generates a new `SummarizedExperiment` object with updated gene annotations and TPM values.
}
