% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_pathways.R
\name{score_pathways}
\alias{score_pathways}
\title{Score Pathways Using Gene Expression Data}
\usage{
score_pathways(exp_data, pathways, scoring_method = "gsva", verbose = TRUE)
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing normalized gene expression data.}

\item{pathways}{A data frame containing pathways with columns `gene_id`, `gene_symbol`, and `pathway`.}

\item{scoring_method}{The method used to score pathways. Options are `"gsva"`, `"ssgsea"`, `"plage"`, or `"zscore"`. Default is `"gsva"`.}

\item{verbose}{Logical. If TRUE, progress messages will be printed. Default is TRUE.}
}
\value{
A data frame containing the scores for the pathways.
}
\description{
This function scores pathways based on gene expression data using various scoring methods such as GSVA, ssGSEA, plage, and zscore.
}
