% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_progeny.R
\name{score_progeny}
\alias{score_progeny}
\title{Score PROGENy Pathways for Gene Expression Data}
\usage{
score_progeny(exp_data, species = "Homo sapiens")
}
\arguments{
\item{exp_data}{A `SummarizedExperiment` object containing normalized gene expression data.}

\item{species}{The species for which to score the pathways. Defaults to "Homo sapiens" ("Human").}
}
\value{
A data frame containing the PROGENy pathway scores for each sample.
}
\description{
This function calculates pathway scores for gene expression data using the PROGENy algorithm.
PROGENy is an algorithm that predicts pathway activity based on gene expression data.
}
